# Fix: rewritten and efficient geo map query