--- a/app/repository/sql/world_map.sql
+++ b/app/repository/sql/world_map.sql
@@ -1,31 +1,33 @@
+-- noqa: disable=all
 -- Filter to published families first
 WITH published_families AS (
-    SELECT f.import_id, f.family_category
-    FROM family AS f
+    SELECT f.import_id, f.family_category 
+    FROM family f
     WHERE EXISTS (
-        SELECT 1
-        FROM family_document AS fd
-        WHERE fd.family_import_id = f.import_id
+        SELECT 1 
+        FROM family_document fd 
+        WHERE fd.family_import_id = f.import_id 
         AND fd.document_status = 'PUBLISHED'
     )
 ),
-
+-- And count how many families are in each geography, filtered by allowed corpuses
 geography_counts AS (
-    SELECT
+    SELECT 
         g.id AS geography_id,
         g.display_value,
         g.slug,
         g.value,
         pf.family_category,
         COUNT(*) AS records_count
-    FROM geography AS g
-    INNER JOIN family_geography AS fg ON g.id = fg.geography_id
-    INNER JOIN published_families AS pf ON fg.family_import_id = pf.import_id
-    INNER JOIN family_corpus AS fc ON pf.import_id = fc.family_import_id
-    WHERE fc.corpus_import_id = ANY(:allowed_corpora_ids)
+    FROM geography g
+    JOIN family_geography fg ON g.id = fg.geography_id
+    JOIN published_families pf ON fg.family_import_id = pf.import_id
+    JOIN family_corpus fc ON pf.import_id = fc.family_import_id
+    WHERE fc.corpus_import_id =  ANY(:allowed_corpora_ids)
     GROUP BY g.id, g.display_value, g.slug, g.value, pf.family_category
 ),
-
+-- This builds a grid where each row is a unique geography/family_category combination
+-- This ensures even when a geography has no families in a category, it will still be included
 category_matrix AS (
     SELECT DISTINCT
         g.id AS geography_id,
@@ -33,13 +35,15 @@ category_matrix AS (
         g.slug,
         g.value,
         f.family_category
-    FROM geography AS g
-    CROSS JOIN f
-),
-
-f AS (SELECT DISTINCT family_category FROM family)
-
-SELECT
+    FROM geography g
+    CROSS JOIN (SELECT DISTINCT family_category FROM family) f
+)
+--Now, we have filtered families that are published;
+-- from that we build a count of how many families are in each geography/family_category combination
+-- and we have a grid of all possible geography/family_category combinations for fast lookup
+-- SO if we left join the geog counts with the matrix, 
+-- we get an output that matches the old world map query without all the extra cruft
+SELECT 
     cm.display_value,
     cm.slug,
     cm.value,
@@ -47,9 +51,9 @@ SELECT
         cm.family_category,
         COALESCE(gc.records_count, 0)
     ) AS counts
-FROM category_matrix AS cm
-LEFT JOIN geography_counts AS gc ON
-    cm.geography_id = gc.geography_id
-    AND cm.family_category = gc.family_category
+FROM category_matrix cm
+LEFT JOIN geography_counts gc ON 
+    cm.geography_id = gc.geography_id AND 
+    cm.family_category = gc.family_category
 GROUP BY cm.display_value, cm.slug, cm.value
 ORDER BY cm.display_value;