# Description

This PR throws away the old geography/map query and introduces a new one that runs in THREE PERCENT of the original query time. THREE PERCENT. 

Anyway, after the fun of the incident I thought it might be a lack of indexing on the key columns in the query. @katybaulch very kindly added those, but I'm sad to report that investigating the performance post-indexing showed query times going from ~12s to ~38s, which if I remember my maths correctly is going in the wrong direction. 

Looking at the query planner, it looks like the indices shifted the query planning to an even more inefficient set of operations. Given how many repeated bits of logic, confusing implicitly materialised views, and so on the original query had, I decided it would probably be easier and quicker to re-write it from scratch. So here we go. 

I hate CASE statements (it's a declarative language lets not shoehorn procedural constructs in), so wanted to get rid of those. I also rather dislike subqueries inside CASE statements, you never know what the query planner will make of them. We do two massive joins at the end of the query on the geographies which are crazy expensive, and it's super hard to tell if we've filtered the DB down to manageable levels or not. 

Instead, I've approached as:
- Filter (ONCE) to published documents
- Use filtered documents to generate document counts 
- Use a cunning trick to build out combinatorial lookup table for all geog/category pairs
- Join that all together to match previous output format.

All investigation using full set of corpus IDs and production database dump taken this morning. 
- Original query on unindexed data:  12s
- Original query on indexed data: 38s
- Updated query in this PR: 0.421s

## Proposed version

Please select the option below that is most relevant from the list below. This
will be used to generate the next tag version name during auto-tagging.

- [ ] Skip auto-tagging
- [x] Patch
- [ ] Minor version
- [ ] Major version

Visit the [Semver website](https://semver.org/#summary) to understand the
difference between `MAJOR`, `MINOR`, and `PATCH` versions.

Notes:

- If none of these options are selected, auto-tagging will fail
- Where multiple options are selected, the most senior option ticked will be
  used -- e.g. Major > Minor > Patch
- If you are selecting the version in the list above using the textbox, make
  sure your selected option is marked `[x]` with no spaces in between the
  brackets and the `x`

## Type of change

Please select the option(s) below that are most relevant:

- [ ] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] GitHub workflow update
- [ ] Documentation update
- [x] Refactor legacy code
- [ ] Dependency update

## How Has This Been Tested?

Please describe the tests that you added to verify your changes.


I've tested by diff'ing outputs of old query and new query on two different DB dumps which show zero diferences.
All tests passing locally. 

THAT SAID I still have a lingering anxiety that the logic is slightly different or I've caused regression somewhere, so some form of inquiry by colleagues with working brains is much appreciated. 

## Reviewer Checklist

- [ ] **DB_CLIENT DEPENDENCY IS ON THE LATEST VERSION**
- [] The PR represents a single feature (small driveby fixes are also ok)
- [ ] The PR includes tests that are sufficient for the level of risk
- [ ] The code is sufficiently commented, particularly in hard-to-understand areas
- [ ] Any required documentation updates have been made
- [ ] Any TODOs added are captured in future tickets
- [ ] No FIXMEs remain