@henry-cpr 2025-05-12
> Looks good, reads well! Any thoughts on testing this for release?

I was hoping for a steer to be honest. I'm still not comfortably idiomatic with navi-backend.  Once I've got CI tests working (which work locally) I think it might need to be deploy on staging and test numbers? I've diffed outputs of old query & new query showing zero difference, but its a big change so I'm loathe to call it a day there. 

I guess it's a spectrum from:
- Deploy to staging and manually have a look
to
- set up a bunch of fixtures with very disparate data distributions and run old and new queries and see if any output diffs.

Or- any magic anyone else thinks of as my brain is very low function today!