@henry-cpr 2025-05-12
NB: trunk fmt changed the semantics of the query, making tests fail in CI that hadn't on local. 

Please challenge if preferred response, but i have disabled linting on the world map query. It's just to tide us over to the geography service...