@jamesgorrie 2025-05-12
@henry-cpr re:testing - @odrakes-cpr had added some timing metrics to our endpoints:
https://github.com/climatepolicyradar/navigator-admin-backend/pull/287

Maybe
- load test `main` on `staging`
- deploy this to `staging`
- load test this on `staging`

Compare numbers?